requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.9.2
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
execnet==2.1.1
factory_boy==3.3.1
Faker==37.4.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
pytest-xdist==3.6.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==75.6.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
django-ckeditor==6.7.3
